optimized a-star heuristic (int)

+++ src/cpp/astar.h
@@ -9,7 +9,7 @@
 
 struct Node {
     int x, y;
-    double gCost, hCost, fCost;
+    int gCost, hCost, fCost;
     Node* parent;
 
     Node() : x(0), y(0), gCost(0), hCost(0), fCost(0), parent(nullptr) {}
@@ -24,7 +24,7 @@ public:
 
 
 private:
-    static double heuristic(const Node& a, const Node& b);
+    static int heuristic(const Node& a, const Node& b);
     static std::vector<Node> getNeighbors(const Node& node, const std::vector<std::vector<bool>>& map);
     static std::vector<Node> reconstructPath(Node* endNode);
     private:
